Introduce AreaOfInterest Download and Refactor Download+Transpose
Introduce a list of areas.
Download the Border Area for all areas but also one per area (which we can later use to enrich the data)
Refactor the Download and Transpose methods into utils.
Filter geometry.type===Point from the Areas with a callback.
Add the "created at" Date (Date of download and transpose)