ZES: Add area key to properties via callback
This introduces logic to enhance our lane data with information of areas. Does the lane intersect (include or cut into) with the given area. The data is save in a true/false pattern per area name and additionally as a list of area values (string).

With this we can filter data based on districts.