ZES: Change download area
This area is now a lot bigger. Its based on a buffer of 10km from the center of the place-nodes of the 5 main locations. Docs for that part will follow.

This also moves some of the general filter logic to the overpass query to keep the file size down. I kept it in the filter as well since the overpass query might change in the future.