ZES: Rework highway data download, refactor property enhancement
Rework the way we add the `FMC:length` to the GeoJSON: Before, we did it by  iterating over all ways to add the length based on a huge array.

Now we have the length as part of the properties right from overpass, BUT we need to manually create our GeoJSON, since osmtogeojson does not support the output of overpass convert.

(For some reasons, we don't need to cut the results to a bbox anymore after switching to convert.)

Related refactorings:
- Add the custom property links (OSM + Mapillary) up front when we create our main GeoJson
- Rename the custom category property
- Disable the addLengthFromOverpassStats method/file since we don't use it anymore